requirements.txt
absl-py==1.0.0; python_version >= "3.6"
alembic==1.7.7; python_version >= "3.7"
antlr4-python3-runtime==4.8; python_version >= "3.6"
astunparse==1.6.3
cachetools==5.0.0; python_version >= "3.7" and python_version < "4.0" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
certifi==2021.10.8; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
charset-normalizer==2.0.12; python_full_version >= "3.6.0" and python_version >= "3.7"
click==8.1.2; python_version >= "3.7"
cloudpickle==2.0.0; python_version >= "3.7"
colorama==0.4.4; python_version >= "3.7" and python_full_version < "3.0.0" and platform_system == "Windows" or python_full_version >= "3.5.0" and platform_system == "Windows" and python_version >= "3.7"
databricks-cli==0.16.6; python_version >= "3.7"
docker==5.0.3; python_version >= "3.7"
entrypoints==0.4; python_version >= "3.7"
flask==2.1.1; python_version >= "3.7"
flatbuffers==2.0
gast==0.4.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
gitdb==4.0.9; python_version >= "3.7"
gitpython==3.1.27; python_version >= "3.7"
google-auth-oauthlib==0.4.6; python_version >= "3.6"
google-auth==2.6.5; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
google-pasta==0.2.0
greenlet==1.1.2; python_version >= "3" and python_full_version < "3.0.0" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") or python_version >= "3" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7") and python_full_version >= "3.5.0"
grpcio==1.45.0; python_version >= "3.6"
gunicorn==20.1.0; platform_system != "Windows" and python_version >= "3.7"
h5py==3.6.0; python_version >= "3.7"
hydra-core==1.1.2
idna==3.3; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
imageio==2.16.1; python_version >= "3.7"
importlib-metadata==4.11.3; python_version < "3.9" and python_version >= "3.7"
importlib-resources==5.2.3; python_version >= "3.7" and python_version < "3.9"
itsdangerous==2.1.2; python_version >= "3.7"
jinja2==3.1.1; python_version >= "3.7"
joblib==1.1.0; python_version >= "3.7"
keras-preprocessing==1.1.2
keras==2.7.0
libclang==13.0.0
llvmlite==0.36.0; python_version >= "3.6" and python_version < "3.10"
mako==1.2.0; python_version >= "3.7"
markdown==3.3.6; python_version >= "3.6"
markupsafe==2.1.1; python_version >= "3.7"
mlflow==1.24.0; python_version >= "3.7"
networkx==2.7.1; python_version >= "3.8"
numba==0.53.1; python_version >= "3.6" and python_version < "3.10"
numpy==1.22.3
oauthlib==3.2.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.7"
omegaconf==2.1.2; python_version >= "3.6"
opt-einsum==3.3.0; python_version >= "3.5"
packaging==21.3; python_version >= "3.7"
pandas==1.4.2; python_version >= "3.8"
pillow==9.1.0; python_version >= "3.7"
prometheus-client==0.13.1; python_version >= "3.7"
prometheus-flask-exporter==0.19.0; python_version >= "3.7"
protobuf==3.20.0; python_version >= "3.7"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6"
pyasn1==0.4.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6") or python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
pyjwt==2.3.0; python_version >= "3.7"
pyparsing==3.0.7; python_version >= "3.7"
python-dateutil==2.8.2; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.8"
pytz==2022.1; python_version >= "3.8"
pywavelets==1.3.0; python_version >= "3.7"
pywin32==227; sys_platform == "win32" and python_version >= "3.7"
pyyaml==6.0; python_version >= "3.7"
querystring-parser==1.2.4; python_version >= "3.7"
requests-oauthlib==1.3.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
requests==2.27.1; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.6")
scikit-image==0.19.2; python_version >= "3.7"
scikit-learn==1.0.2; python_version >= "3.7"
scipy==1.8.0; python_version >= "3.8" and python_version < "3.11"
segyio==1.9.9
shap==0.40.0
six==1.16.0; python_version >= "3.8" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.8"
slicer==0.0.7; python_version >= "3.6"
smmap==5.0.0; python_version >= "3.7"
sqlalchemy==1.4.34; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
sqlparse==0.4.2; python_version >= "3.7"
tabulate==0.8.9; python_version >= "3.7"
tensorboard-data-server==0.6.1; python_version >= "3.6"
tensorboard-plugin-wit==1.8.1; python_version >= "3.6"
tensorboard==2.8.0; python_version >= "3.6"
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.24.0; python_version >= "3.7" and python_version < "3.11"
tensorflow==2.7.0
termcolor==1.1.0
threadpoolctl==3.1.0; python_version >= "3.7"
tifffile==2022.3.25; python_version >= "3.8"
tqdm==4.63.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.4.0"
typing-extensions==4.1.1; python_version >= "3.6"
urllib3==1.26.9; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version < "4" and python_version >= "3.7"
waitress==2.1.1; platform_system == "Windows" and python_version >= "3.7" and python_full_version >= "3.7.0"
websocket-client==1.3.2; python_version >= "3.7"
werkzeug==2.1.1; python_version >= "3.7"
wrapt==1.14.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
zipp==3.7.0; python_version < "3.9" and python_version >= "3.7"
